+++ config/services.php
@@ -61,4 +61,10 @@ return [
         'mode' => env('PAYPAL_MODE', 'sandbox'),
     ],
 
+    'google' => [
+        'client_id' => env('GOOGLE_CLIENT_ID', env('GOOGLE_OAUTH_CLIENT_ID')),
+        'client_secret' => env('GOOGLE_CLIENT_SECRET', env('GOOGLE_OAUTH_CLIENT_SECRET')),
+        'redirect' => env('GOOGLE_REDIRECT_URI'),
+    ],
+
 ];